+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add id mapped dataset class
+- Added id mapped dataset class
+- Added canonical name property 
 
 ## [0.1.1] - 2023-03-27